tag for release (TrigFTKSim-00-11-71)
	* Tag as TrigFTKSim-00-11-71 for release

2016-05-04 Jahred Adelman
	* With pbank format and nsub=1, don't do double merge, which saves time
	* Tag as TrigFTKSim-00-11-69

2016-05-02 Jahred Adelman
	* Updates for 2016 barrel only banks
	* Tag as TrigFTKSim-00-11-68

2016-04-24 Jahred Adelman
	* Update CMakeLists.txt to add all ftk configuration files
	* Tag as TrigFTKSim-00-11-65

2016-04-20 Thor Taylor
	* update to RTT code to use new FTKIP. Some placeholder jobs also
	added.
	* Tag as TrigFTKSim-00-11-64

2016-03-22 Jahred Adelman
...
(Long ChangeLog diff - truncated)


Former-commit-id: 387b005e2063657b1db03e0d0e36fec9d54029c0

+++ Trigger/TrigFTK/TrigFTKSim/TrigFTKSim/MultiTruth.h
@@ -5,7 +5,7 @@
 #ifndef WRAP_MULTITRUTH_H
 #define WRAP_MULTITRUTH_H
 
-// MultiTruth 
+// MultiTruth
 // ================================================================
 // code to match clusters and tracks to GEANT charge deposition information
 // ================================================================
@@ -24,7 +24,7 @@
 // class which represents the relative contributions of one or more
 // truth particles (identified by a unique "barcode") to some
 // measurement. for example, a single silicon pixel may record the
-// summed ionization from two different charged partilces passing
+// summed ionization from two different charged particles passing
 // through it during the same event. one particle may (and likely
 // will) contribute more charge than another, so the class stores a
 // weight for each barcode which can be used to incorporate this
@@ -43,7 +43,7 @@ class
 MultiTruth : public TObject
 {
 public:
-  typedef std::pair<long int,long int> Barcode;
+  typedef std::pair<long int,long int> Barcode; // = (event index, barcode)
   typedef float Weight;
   struct MaxAccumulator {
     MultiTruth operator()( const MultiTruth& result , const MultiTruth& a ) const { return result.maximize(a); }
@@ -54,11 +54,11 @@ public:
 protected:
   typedef std::map<Barcode,Weight> TruthMap;
   struct TruthMapWeightLt {
-    bool operator()( const TruthMap::value_type& a , const TruthMap::value_type& b ) const { 
+    bool operator()( const TruthMap::value_type& a , const TruthMap::value_type& b ) const {
       const bool a_info = a.first.first!=-1 && a.first.second!=-1;
       const bool b_info = b.first.first!=-1 && b.first.second!=-1;
       if( a_info && b_info ) {
-        return( a.second < b.second ); 
+        return( a.second < b.second );
       } else {
         return b_info;
       }
@@ -79,7 +79,7 @@ public:
   // tricky: overloading derived class assignment operator
   MultiTruth& operator=(const MultiTruth& cpy) {
 #ifdef PROTECT_SA // speedup
-    if(this != &cpy) 
+    if(this != &cpy)
 #endif
       {
 	this->TObject::operator=(cpy);
@@ -102,6 +102,7 @@ public:
       i->second += weight;
     }
   }
+
   void add(const MultiTruth& rval) {
     for( TruthMap::const_iterator i=rval._truth.begin(), f=rval._truth.end(); i!=f; ++i ) {
       TruthMap::iterator j=_truth.find( i->first );
@@ -112,6 +113,7 @@ public:
       }
     }
   }
+
   void assign_equal_normalization() {
     if( _truth.empty() ) { return; }
     unsigned int ncodes = _truth.size();
@@ -119,6 +121,7 @@ public:
       i->second = 1./ncodes;
     }
   }
+
   void maximize(const MultiTruth::Barcode& code,const MultiTruth::Weight& weight) {
     // if this doesn't have any truth info, assign maxweight_barcode to be the truth.
     // otherwise, this should only have the barcode of the largest weight seen so far;
@@ -133,6 +136,7 @@ public:
       }
     }
   }
+
   void maximize(const MultiTruth& rval) {
     // if rval has no truth info, do nothing.
     if( rval._truth.empty() ) { return; }
@@ -147,19 +151,23 @@ public:
     }
     maximize(maxweight_barcode,maxweight);
   }
+
   // add and mult for std::accumulate
   const MultiTruth add(const MultiTruth& rval) const {
     MultiTruth result(*this);
     result.add(rval);
     return( result );
   }
+
   const MultiTruth maximize(const MultiTruth& rval) const {
     MultiTruth result(*this);
     result.maximize(rval);
     return( result );
   }
+
   // matching probability definition and maximization logic
   inline MultiTruth::Weight total_weight() const { return std::accumulate( _truth.begin() , _truth.end() , 0. , TruthMapWeightAcc() ); }
+
   inline MultiTruth::Weight weight(const MultiTruth::Barcode& code) const {
     if( _truth.empty() ) { return 0.; }
     TruthMap::const_iterator i=_truth.find(code);
@@ -167,6 +175,7 @@ public:
     const MultiTruth::Weight tot = total_weight();
     return( i->second / tot );
   }
+
   inline bool best( MultiTruth::Barcode& code , MultiTruth::Weight& weight ) const {
     if( _truth.empty() ) { return false; }
     TruthMap::const_iterator i = std::max_element( _truth.begin() , _truth.end() , TruthMapWeightLt() );
@@ -175,8 +184,9 @@ public:
     weight = tot>0. ? i->second/tot : 0.;
     return true;
   }
+
   inline unsigned int multiplicity() const { return _truth.size(); }
-  unsigned int best_barcode() const;
+  long int best_barcode() const;
   // write contents of this object to stdout
   void display() const;
   // this is a TObject to be stored in the TrigFTKSim ROOT output streams
